Moved Exp code to Pokémon Core module
