+++ Source/PokemonCore/Public/Pokemon/Exp/Fluctuating.h
@@ -16,10 +16,13 @@
 #include "CoreMinimal.h"
 #include "GrowthRate.h"
 
-/**
- * Represents the Parabolic Exp. Growth Scheme
- */
-class POKEMONDATA_API FParabolic : public IGrowthRate {
-public:
-	int ExpForLevel(int Level) const override;
-};
+namespace Exp {
+	/**
+	* Represents the Fluctuating Exp. Growth Scheme
+	*/
+	class POKEMONCORE_API FFluctuating : public IGrowthRate {
+
+	public:
+		int ExpForLevel(int Level) const override;
+	};
+}
