+++ Source/PokemonCore/Public/Pokemon/Exp/GrowthRate.h
@@ -1,4 +1,4 @@
-//====================================================================================================================
+﻿//====================================================================================================================
 // ** Unreal Pokémon created by Retro & Chill
 //--------------------------------------------------------------------------------------------------------------------
 // This project is intended as a means of learning more about how a game like Pokémon works by creating a framework
@@ -11,13 +11,25 @@
 // TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 //====================================================================================================================
-#include "Exp/Fast.h"
+#pragma once
 
-int FFast::ExpForLevel(int Level) const {
-	check(Level > 0);
+#include "CoreMinimal.h"
 
-	if (Level == 1)
-		return 0;
 
-	return std::pow(Level, 3) * 4 / 5;
-}
+namespace Exp {
+	/**
+	 * Interface to represent various Exp. grow rates
+	 */
+	class POKEMONCORE_API IGrowthRate {
+
+	public:
+		virtual ~IGrowthRate() = default;
+	
+		/**
+		 * Gets the amount of Exp that is required to reach a certain level 
+		 * @param Level The level in to calculate the Exp. for
+		 * @return The amount of Exp. required to level up
+		 */
+		virtual int ExpForLevel(int Level) const = 0;
+	};
+};
