+++ Source/PokemonCore/Tests/Pokemon/Exp/GrowthRateErraticTest.cpp
@@ -15,7 +15,7 @@
 
 #include <array>
 
-#include "Exp/Erratic.h"
+#include "Pokemon/Exp/Erratic.h"
 
 constexpr std::array GErraticGrowthTable = {
 	-1,
@@ -72,7 +72,7 @@ void GrowthRateErraticTest::GetTests(TArray<FString>& OutBeautifiedNames, TArray
 bool GrowthRateErraticTest::RunTest(const FString& Parameters) {
 	const int32 Level = FCString::Atoi(GetData(Parameters));
 
-	FErratic GrowthRate;
+	Exp::FErratic GrowthRate;
 	return TestEqual("The amount of Exp. required to level up should match the expected value!",
 	                 GrowthRate.ExpForLevel(Level), GErraticGrowthTable[Level]);
 }
