+++ Source/PokemonCore/Tests/Pokemon/Exp/GrowthRateMediumTest.cpp
@@ -11,46 +11,47 @@
 // TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 //====================================================================================================================
-#include "Exp/Parabolic.h"
 #include "Misc/AutomationTest.h"
 
 #include <array>
 
-constexpr std::array GParabolicGrowthTable = {
+#include "Pokemon/Exp/Medium.h"
+
+constexpr std::array GMediumGrowthTable = {
 	-1,
-	0, 9, 57, 96, 135, 179, 236, 314, 419, 560,
-	742, 973, 1261, 1612, 2035, 2535, 3120, 3798, 4575, 5460,
-	6458, 7577, 8825, 10208, 11735, 13411, 15244, 17242, 19411, 21760,
-	24294, 27021, 29949, 33084, 36435, 40007, 43808, 47846, 52127, 56660,
-	61450, 66505, 71833, 77440, 83335, 89523, 96012, 102810, 109923, 117360,
-	125126, 133229, 141677, 150476, 159635, 169159, 179056, 189334, 199999, 211060,
-	222522, 234393, 246681, 259392, 272535, 286115, 300140, 314618, 329555, 344960,
-	360838, 377197, 394045, 411388, 429235, 447591, 466464, 485862, 505791, 526260,
-	547274, 568841, 590969, 613664, 636935, 660787, 685228, 710266, 735907, 762160,
-	789030, 816525, 844653, 873420, 902835, 932903, 963632, 995030, 1027103, 1059860
+	0, 8, 27, 64, 125, 216, 343, 512, 729, 1000,
+	1331, 1728, 2197, 2744, 3375, 4096, 4913, 5832, 6859, 8000,
+	9261, 10648, 12167, 13824, 15625, 17576, 19683, 21952, 24389, 27000,
+	29791, 32768, 35937, 39304, 42875, 46656, 50653, 54872, 59319, 64000,
+	68921, 74088, 79507, 85184, 91125, 97336, 103823, 110592, 117649, 125000,
+	132651, 140608, 148877, 157464, 166375, 175616, 185193, 195112, 205379, 216000,
+	226981, 238328, 250047, 262144, 274625, 287496, 300763, 314432, 328509, 343000,
+	357911, 373248, 389017, 405224, 421875, 438976, 456533, 474552, 493039, 512000,
+	531441, 551368, 571787, 592704, 614125, 636056, 658503, 681472, 704969, 729000,
+	753571, 778688, 804357, 830584, 857375, 884736, 912673, 941192, 970299, 1000000
 };
 
-IMPLEMENT_COMPLEX_AUTOMATION_TEST(GrowthRateParabolicTest, "Tests.Exp.GrowthRateParabolicTest",
+IMPLEMENT_COMPLEX_AUTOMATION_TEST(GrowthRateMediumTest, "Tests.Exp.GrowthRateMediumTest",
                                   EAutomationTestFlags::EditorContext | EAutomationTestFlags::EngineFilter)
 
-void GrowthRateParabolicTest::GetTests(TArray<FString>& OutBeautifiedNames, TArray<FString>& OutTestCommands) const {
-	OutBeautifiedNames.Add("Tests.Exp.GrowthRateParabolicTest: Level = 1");
+void GrowthRateMediumTest::GetTests(TArray<FString>& OutBeautifiedNames, TArray<FString>& OutTestCommands) const {
+	OutBeautifiedNames.Add("Tests.Exp.GrowthRateMediumTest: Level = 1");
 	OutTestCommands.Add("1");
 
 	for (int i = 0; i < 3; i++) {
-		auto Fmt = FString::Printf(TEXT("Tests.Exp.GrowthRateParabolicTest: Level = 2-99 (Sample %d)"), i + 1);
+		auto Fmt = FString::Printf(TEXT("Tests.Exp.GrowthRateMediumTest: Level = 2-99 (Sample %d)"), i + 1);
 		OutBeautifiedNames.Add(Fmt);
 		OutTestCommands.Add(FString::FromInt(FMath::RandRange(2, 99)));
 	}
 
-	OutBeautifiedNames.Add("Tests.Exp.GrowthRateParabolicTest: Level = 100");
+	OutBeautifiedNames.Add("Tests.Exp.GrowthRateMediumTest: Level = 100");
 	OutTestCommands.Add("100");
 }
 
-bool GrowthRateParabolicTest::RunTest(const FString& Parameters) {
+bool GrowthRateMediumTest::RunTest(const FString& Parameters) {
 	const int32 Level = FCString::Atoi(GetData(Parameters));
 
-	FParabolic GrowthRate;
+	Exp::FMedium GrowthRate;
 	return TestEqual("The amount of Exp. required to level up should match the expected value!",
-	                 GrowthRate.ExpForLevel(Level), GParabolicGrowthTable[Level]);
+	                 GrowthRate.ExpForLevel(Level), GMediumGrowthTable[Level]);
 }
